Migrates to ov script in cli module
