+++ cli/__init__.py
@@ -1,0 +1,3 @@
+from .cli import app
+
+__all__ = ['app']
